Adição de verificação de erros
