+++ Dockerfile
@@ -5,4 +5,4 @@ RUN pip install --no-cache-dir -r requirements.txt
 EXPOSE 5000
 COPY . .
 ENTRYPOINT ["/usr/local/bin/flask"]
-CMD ["--app", "api", "run", "--host=0.0.0.0"]
+CMD ["run", "--host=0.0.0.0"]
