+++ app.py
@@ -1,0 +1,142 @@
+from flask import Flask, request, jsonify, abort
+from utils.exceptions import ValueNotInformed, ValueNotValid
+from modules.system import System
+import numpy as np
+import re
+
+app = Flask(__name__)
+
+
+def format_root(root):
+    root = re.sub('[iIjJ]', 'j', root)
+    root = re.sub('([-+])j$', '\g<1>1j', root)
+    root = re.sub('^j$', '1j', root)
+    return root
+
+
+def get_poly(str_input, type, field):
+    if type == 'poly':
+        try:
+            poly = list(map(float, str_input.split()))
+            return poly
+        except:
+            raise ValueNotValid(field)
+
+    elif type == 'roots':
+        try:
+            roots = list(map(format_root, str_input.split()))
+            roots = list(map(complex, roots))
+        except:
+            raise ValueNotValid(field)
+
+        poly = [1]
+        for r in roots:
+            poly = np.convolve(poly, [1, -r])
+        poly = np.round(poly, 4)
+
+        if any(np.imag(poly)):
+            raise Exception(
+                f'O polinômio resultante do {field} possui conficiêntes complexos.')
+
+        return np.real(poly)
+
+    else:
+        raise Exception(f'Tipo do {field.lower()} não é valido')
+
+
+def process_system(system_data: dict) -> System:
+
+    if 'num' not in system_data:
+        raise ValueNotInformed(field="Numerador do sistema")
+    if 'den' not in system_data:
+        raise ValueNotInformed(field="Denominador do sistema")
+    if 'num_type' not in system_data:
+        raise ValueNotInformed(field="Tipo de numerador do sistema")
+    if 'den_type' not in system_data:
+        raise ValueNotInformed(field="Tipo de denominador do sistema")
+
+    num = get_poly(system_data['num'], system_data.get(
+        'num_type'), "Numerador do sistema")
+    den = get_poly(system_data['den'], system_data.get(
+        'den_type'), "Denominador do sistema")
+
+    gain = system_data.get('gain', 1)
+    feedback = system_data.get('feedback', False)
+
+    sys = System(num, den, feedback, gain)
+
+    if 'comp' in system_data:
+        comp_data = system_data['comp']
+
+        if 'num' not in system_data:
+            raise ValueNotInformed(field="Numerador do compensador")
+        if 'den' not in system_data:
+            raise ValueNotInformed(field="Denominador do sistema")
+        if 'num_type' not in system_data:
+            raise ValueNotInformed(field="Tipo de numerador do compensador")
+        if 'den_type' not in system_data:
+            raise ValueNotInformed(field="Tipo de denominador do compensador")
+
+        num = get_poly(comp_data['num'], comp_data.get(
+            'num_type'), field="Numerador do compensador")
+        den = get_poly(comp_data['den'], comp_data.get(
+            'den_type'), field="Denominador do sistema")
+        gain = comp_data.get('gain', 1)
+        sys.conf_comp(num, den, gain)
+
+    if 'pid' in system_data:
+        pid_data = system_data['pid']
+        kp = pid_data.get('kp', 0)
+        kd = pid_data.get('kd', 0)
+        ki = pid_data.get('ki', 0)
+
+        if kp or kd or ki:
+            sys.conf_pid(kp, ki, kd)
+
+    return sys
+
+
+def process_simulations(data: dict, sys: System) -> dict:
+    results = {'plots': {}, 'values': {}}
+
+    if 'plots' in data:
+        plots = data['plots']
+        if 'step_response' in plots:
+            results['plots']['step_response'] = sys.step_response()
+        if 'pzmap' in plots:
+            results['plots']['pzmap'] = sys.pzmap()
+        if 'rlocus' in plots:
+            results['plots']['pzmap'] = sys.rlocus()
+
+    if 'values' in data:
+        values = data['values']
+        if 'system' in values:
+            results['values']['system'] = str(sys)
+        if 'zeros' in values:
+            results['values']['zeros'] = sys.zeros()
+        if 'poles' in values:
+            results['values']['poles'] = sys.poles()
+
+    return results
+
+
+def process_json(data: dict) -> dict:
+    sys = process_system(data['system'])
+    results = process_simulations(data, sys)
+    return results
+
+
+@app.route('/', methods=['post'])
+def index():
+    try:
+        data = request.get_json()
+        if 'system' in data:
+            results = process_json(data)
+    except Exception as error:
+        abort(400, description=str(error))
+    return jsonify(results=results), 200
+
+
+@app.errorhandler(400)
+def json_error(e):
+    return jsonify(Error=e.description), 400
